+++ src/css/style.css
@@ -12,4 +12,3 @@ th, td {
   padding: 10px 20px 10px 20px;
   text-align: center;
 }
-
